refactor: 동아리 QNA 도메인을 하위 패키지로 분리
