+++ src/main/java/in/koreatech/koin/domain/club/qna/controller/ClubQnaApi.java
@@ -1,0 +1,94 @@
+package in.koreatech.koin.domain.club.qna.controller;
+
+import in.koreatech.koin.domain.club.qna.dto.request.ClubQnaCreateRequest;
+import in.koreatech.koin.domain.club.qna.dto.response.ClubQnasResponse;
+import in.koreatech.koin.global.auth.Auth;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import static in.koreatech.koin.domain.user.model.UserType.STUDENT;
+import static io.swagger.v3.oas.annotations.enums.ParameterIn.PATH;
+
+@Tag(name = "(Normal) Club QNA: 동아리 QNA", description = "동아리 QNA 정보를 관리한다")
+@RequestMapping("/clubs/{clubId}/qna")
+public interface ClubQnaApi {
+
+    @Operation(
+        summary = "특정 동아리의 모든 QNA를 조회한다",
+        description = """
+            - authorId 확인하여 작성자 본인인 경우 삭제 버튼(x) 표시.
+            - 닉네임은 존재 시 그대로 반환되며, 없는 경우 student의 익명 닉네임으로 반환.
+            - 트리 구조는 대댓글 형태로 재귀적으로 구성됩니다.
+            
+            ```java
+            예시
+            댓글 1
+            ├── 댓글 1-1
+            │   ├── 댓글 1-1-1
+            │   │   └── 댓글 1-1-1-1
+            │   └── 댓글 1-1-2
+            ├── 댓글 1-2
+            └── 댓글 1-3
+            
+            댓글 2
+            └── 댓글 2-1
+                └── 댓글 2-1-1
+            ```
+            """
+    )
+    @GetMapping
+    ResponseEntity<ClubQnasResponse> getQnas(
+        @Parameter(in = PATH) @PathVariable Integer clubId
+    );
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "201"),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
+        }
+    )
+    @Operation(summary = "특정 동아리의 QNA를 생성한다",
+        description = """
+            사용자의 경우 질문만 가능, 관리자의 경우 답변만 가능
+            parentId를 null 요청 시 질문, 질문 QNA의 id를 넣어서 요청하면 답변 형식으로 생성
+            """
+    )
+    @PostMapping
+    ResponseEntity<Void> createQna(
+        @RequestBody @Valid ClubQnaCreateRequest request,
+        @Parameter(in = PATH) @PathVariable Integer clubId,
+        @Auth(permit = {STUDENT}) Integer studentId
+    );
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "204"),
+            @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
+        }
+    )
+    @Operation(summary = "특정 동아리의 QNA를 삭제한다",
+        description = """
+            - 관리자는 모든 QNA 삭제 가능, 그 외에는 본인의 QNA만 삭제 가능
+            - 부모 QNA(질문 QNA)인 경우, 답변 QNA까지 모두 삭제
+            - 자식 QNA(답변 QNA)인 경우, 답변 QNA만 삭제
+            """)
+    @DeleteMapping("/{qnaId}")
+    ResponseEntity<Void> deleteQna(
+        @Parameter(in = PATH) @PathVariable Integer clubId,
+        @Parameter(in = PATH) @PathVariable Integer qnaId,
+        @Auth(permit = {STUDENT}) Integer studentId
+    );
+}
