+++ src/main/java/in/koreatech/koin/domain/club/qna/dto/response/ClubQnasResponse.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.club.dto.response;
+package in.koreatech.koin.domain.club.qna.dto.response;
 
 import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
@@ -9,7 +9,7 @@ import java.util.List;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
-import in.koreatech.koin.domain.club.model.ClubQna;
+import in.koreatech.koin.domain.club.qna.model.ClubQna;
 import in.koreatech.koin.domain.student.model.Student;
 import io.swagger.v3.oas.annotations.media.Schema;
