+++ src/main/java/in/koreatech/koin/domain/club/qna/exception/ClubQnaNotFoundException.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.club.exception;
+package in.koreatech.koin.domain.club.qna.exception;
 
 import in.koreatech.koin.global.exception.custom.DataNotFoundException;
