+++ src/test/java/in/koreatech/koin/unit/fixture/ClubQnaFixture.java
@@ -1,7 +1,7 @@
 package in.koreatech.koin.unit.fixture;
 
 import in.koreatech.koin.domain.club.model.Club;
-import in.koreatech.koin.domain.club.model.ClubQna;
+import in.koreatech.koin.domain.club.qna.model.ClubQna;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import java.time.LocalDateTime;
